refactor plotting tasks.
