+++ src/policies/task_visualize_work_multipliers.py
@@ -5,7 +5,7 @@ import seaborn as sns
 from sid.colors import get_colors
 
 from src.config import BLD
-from src.simulation.plotting import style_plot
+from src.plotting.plotting import style_plot
 
 plt.rcParams.update(
     {
