+++ src/testing/task_calculate_share_known_cases.py
@@ -7,7 +7,7 @@ import seaborn as sns
 
 from src.config import BLD
 from src.config import SRC
-from src.simulation.plotting import style_plot
+from src.plotting.plotting import style_plot
 from src.testing.shared import get_piecewise_linear_interpolation
