+++ src/testing/task_plot_share_of_educ_participants_with_rapid_test.py
@@ -6,7 +6,7 @@ import pytask
 import seaborn as sns
 
 from src.config import BLD
-from src.simulation.plotting import style_plot
+from src.plotting.plotting import style_plot
 from src.testing.shared import get_piecewise_linear_interpolation
